@@ -16,18 +16,18 @@
    },
    "devDependencies": {
       "@sveltejs/adapter-auto": "^3.0.0",
-      "@sveltejs/kit": "^2.5.4",
-      "@sveltejs/vite-plugin-svelte": "^3.0.0",
+      "@sveltejs/kit": "^2.5.9",
+      "@sveltejs/vite-plugin-svelte": "^3.1.0",
       "@typescript-eslint/eslint-plugin": "^6.0.0",
       "@typescript-eslint/parser": "^6.0.0",
       "autoprefixer": "^10.4.16",
-      "eslint": "^8.28.0",
+      "eslint": "^9.3.0",
       "eslint-config-prettier": "^9.1.0",
-      "eslint-plugin-svelte": "^2.30.0",
+      "eslint-plugin-svelte": "^2.39.0",
       "postcss": "^8.4.32",
       "prettier": "^3.1.1",
       "prettier-plugin-svelte": "^3.1.2",
-      "svelte": "^4.2.7",
+      "svelte": "^5.0.0-next.136",
       "svelte-check": "^3.6.0",
       "svelte-motion": "^0.12.1",
       "tailwindcss": "^3.4.1",
@@ -38,7 +38,7 @@
    },
    "type": "module",
    "dependencies": {
-      "@auth/sveltekit": "^0.14.0",
+      "@auth/sveltekit": "^1.1.0",
       "@fontsource-variable/inter": "^5.0.16",
       "@formkit/auto-animate": "^0.8.1",
       "@microsoft/signalr": "^8.0.0",
