@@ -4,16 +4,16 @@
 <div class="items-center rounded-md border-2 border-muted p-1 hover:border-accent relative">
    <div class="space-y-2 rounded-sm bg-[#ecedef] p-2">
       <div class="space-y-2 rounded-md bg-white p-2 shadow-sm">
-         <div class="h-2 w-[80px] rounded-lg bg-[#ecedef]" />
-         <div class="h-2 w-[100px] rounded-lg bg-[#ecedef]" />
+         <div class="h-2 w-[80px] rounded-lg bg-[#ecedef]"></div>
+         <div class="h-2 w-[100px] rounded-lg bg-[#ecedef]"></div>
       </div>
       <div class="flex items-center space-x-2 rounded-md bg-slate-800 p-2 shadow-sm">
-         <div class="h-4 w-4 rounded-full bg-slate-400" />
-         <div class="h-2 w-[100px] rounded-lg bg-slate-400" />
+         <div class="h-4 w-4 rounded-full bg-slate-400"></div>
+         <div class="h-2 w-[100px] rounded-lg bg-slate-400"></div>
       </div>
       <div class="flex items-center space-x-2 rounded-md bg-slate-800 p-2 shadow-sm">
-         <div class="h-4 w-4 rounded-full bg-slate-400" />
-         <div class="h-2 w-[100px] rounded-lg bg-slate-400" />
+         <div class="h-4 w-4 rounded-full bg-slate-400"></div>
+         <div class="h-2 w-[100px] rounded-lg bg-slate-400"></div>
       </div>
    </div>
 </div>
