@@ -8,4 +8,4 @@
    export { className as class };
 </script>
 
-<div class={cn('animate-pulse rounded-md bg-gray-800', className)} {...$$restProps} />
+<div class={cn('animate-pulse rounded-md bg-gray-800', className)} {...$$restProps}></div>
