@@ -1,36 +1,50 @@
 <script lang="ts">
-   import { page } from '$app/stores';
-   import { afterUpdate } from 'svelte';
-
    import Text from '$lib/design-system/Text.svelte';
    import Heading from '$lib/design-system/Heading.svelte';
-   import { chatState } from '$lib/modules/chat/contexts/chat-context/stores/chat';
 
    import Message from './Message.svelte';
    import Typebar from './Typebar.svelte';
    import ScrollToBottomButton from './ScrollToBottomButton.svelte';
 
-   let messagesContainer: HTMLUListElement;
+   import { chatState } from '$lib/modules/chat/contexts/chat-context/stores/chat';
 
-   $: isFirstAccess = true;
-   $: userId = $page.data.session?.user.id;
-   $: selectedConversationId = $chatState.selectedConversation?.id;
-   $: conversationMessages = $chatState.selectedConversation?.messages || [];
+   import type { ILocalMessage } from '../../../../../../types/message';
 
-   afterUpdate(() => {
-      if (selectedConversationId || !selectedConversationId) {
-         isFirstAccess = true;
+   let messagesContainer: HTMLUListElement | null = $state(null);
+   let firstUnreadMessage: ILocalMessage | null = $state(null);
+
+   let userId = $derived(globalThis.$page.data.session?.user.id);
+   let isFirstAccess = $state(true);
+   let conversationMessages = $derived($chatState.selectedConversation?.messages || []);
+
+   $effect(() => {
+      if (!firstUnreadMessage) {
+         firstUnreadMessage = conversationMessages.find(message => message.read.users.some(user => !user.at)) || null;
       }
+   });
+
+   $effect(() => {
+      $chatState.selectedConversation?.id;
+      // The code above is only to track the selected conversation to the $effect, to
+      // run this function when the user change the selected chat
 
       messagesContainer?.scroll({
          top: messagesContainer.scrollHeight,
          behavior: isFirstAccess ? 'auto' : 'smooth',
       });
+   });
 
-      isFirstAccess = false;
+   $effect(() => {
+      isFirstAccess = true;
+
+      return () => {
+         isFirstAccess = false;
+      };
    });
 </script>
 
+<!-- svelte-ignore non_reactive_update -->
+<!-- svelte-ignore non_reactive_update -->
 <div class="overflow-hidden w-full flex-1 h-full flex flex-col p-1 relative max-w-5xl mx-auto">
    <ul class="flex flex-col gap-2 overflow-auto p-4" bind:this={messagesContainer}>
       {#if !$chatState.selectedConversation?.hasMessagesFetched}
@@ -44,9 +58,10 @@
          </div>
       {:else if $chatState.selectedConversation?.messages}
          {#each conversationMessages as message, i (message.id)}
-            <!-- {#if firstUnreadMessage?.id === message.id && message.author.id !== userId}
-                     <div class="w-full h-[1px] bg-red-500 flex justify-center">Unread messages</div>
-                  {/if} -->
+            {#if firstUnreadMessage?.id === message.id}
+               <div class="w-full h-[1px] bg-red-500 flex justify-center text-white">Unread messages</div>
+            {/if}
+
             <Message {message} previousMessage={conversationMessages?.[i - 1]} messageIndex={i} />
          {/each}
       {/if}
