@@ -17,7 +17,7 @@ export async function getStatusFromMessage(
    { messageId, conversationId }: Request,
    { accessToken }: IServiceOptions
 ): Promise<Response> {
-   const { data } = await api.get<Response>(`/message/status/${conversationId}/${messageId}`, {
+   const { data } = await api.get<Response>(`/messages/status/${conversationId}/${messageId}`, {
       headers: {
          Authorization: `Bearer ${accessToken}`,
       },
